Added backend infrastructure for enchanced publication assessment scheduling. Fixed authorization errors.
