+++ src/main/java/rs/teslaris/core/assessment/repository/EventAssessmentClassificationRepository.java
@@ -1,6 +1,7 @@
 package rs.teslaris.core.assessment.repository;
 
 import java.util.List;
+import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
@@ -13,4 +14,12 @@ public interface EventAssessmentClassificationRepository extends
     @Query("select eac from EventAssessmentClassification eac where " +
         "eac.event.id = :eventId order by eac.timestamp desc")
     List<EventAssessmentClassification> findAssessmentClassificationsForEvent(Integer eventId);
+
+    @Query(
+        "SELECT eac FROM EventAssessmentClassification eac WHERE " +
+            "eac.event.id = :eventId AND " +
+            "eac.commission.id = :commissionId AND " +
+            "eac.classificationYear = :year")
+    Optional<EventAssessmentClassification> findAssessmentClassificationsForEventAndCommissionAndYear(
+        Integer eventId, Integer commissionId, Integer year);
 }
