+++ src/main/java/rs/teslaris/core/assessment/service/interfaces/DocumentAssessmentClassificationService.java
@@ -2,6 +2,7 @@ package rs.teslaris.core.assessment.service.interfaces;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import rs.teslaris.core.assessment.dto.EntityAssessmentClassificationResponseDTO;
@@ -15,5 +16,9 @@ public interface DocumentAssessmentClassificationService {
     void classifyJournalPublication(Integer journalPublicationId);
 
     void scheduleJournalPublicationClassification(LocalDateTime timeToRun,
-                                                  Integer userId, LocalDate fromDate);
+                                                  Integer userId, LocalDate fromDate,
+                                                  Integer commissionId,
+                                                  ArrayList<Integer> authorIds,
+                                                  ArrayList<Integer> orgUnitIds,
+                                                  ArrayList<Integer> journalIds);
 }
