+++ src/main/java/rs/teslaris/core/configuration/SecurityConfiguration.java
@@ -156,6 +156,9 @@ public class SecurityConfiguration {
 
                 // EVENTS
                 .requestMatchers(HttpMethod.GET, "/api/event/{eventId}").permitAll()
+                .requestMatchers(HttpMethod.GET, "/api/conference/{conference}").permitAll()
+                .requestMatchers(HttpMethod.GET, "/api/conference/old-id/{oldConferenceId}")
+                .permitAll()
 
                 // EVENTS RELATION
                 .requestMatchers(HttpMethod.GET, "/api/events-relation/{eventId}").permitAll()
@@ -168,6 +171,8 @@ public class SecurityConfiguration {
 
                 // RESEARCH AREA
                 .requestMatchers(HttpMethod.GET, "/api/research-area").permitAll()
+                .requestMatchers(HttpMethod.GET, "/api/research-area/children/{parentId}")
+                .permitAll()
 
                 // FILE
                 .requestMatchers(HttpMethod.GET, "/api/file/{serverFilename}").permitAll()
@@ -181,15 +186,15 @@ public class SecurityConfiguration {
                 // ASSESSMENT
                 .requestMatchers(HttpMethod.GET, "/api/assessment/document-indicator/{documentId}")
                 .permitAll()
-                .requestMatchers(HttpMethod.GET, "/api/assessment/person-indicator/{documentId}")
+                .requestMatchers(HttpMethod.GET, "/api/assessment/person-indicator/{personId}")
                 .permitAll()
-                .requestMatchers(HttpMethod.GET, "/api/assessment/event-indicator/{documentId}")
-                .permitAll()
-                .requestMatchers(HttpMethod.GET,
-                    "/api/assessment/organisation-unit-indicator/{documentId}")
+                .requestMatchers(HttpMethod.GET, "/api/assessment/event-indicator/{eventId}")
                 .permitAll()
                 .requestMatchers(HttpMethod.GET,
-                    "/api/assessment/publication-series-indicator/{documentId}")
+                    "/api/assessment/organisation-unit-indicator/{organisationUnitId}")
+                .permitAll()
+                .requestMatchers(HttpMethod.GET,
+                    "/api/assessment/publication-series-indicator/{publicationSeriesId}")
                 .permitAll()
 
                 // STATISTICS
