+++ src/main/java/rs/teslaris/dbinitialization/DbInitializer.java
@@ -1,0 +1,315 @@
+package rs.teslaris.dbinitialization;
+
+
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.boot.ApplicationArguments;
+import org.springframework.boot.ApplicationRunner;
+import org.springframework.context.annotation.Profile;
+import org.springframework.core.env.Environment;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
+import rs.teslaris.core.model.commontypes.Country;
+import rs.teslaris.core.model.commontypes.Language;
+import rs.teslaris.core.model.commontypes.LanguageTag;
+import rs.teslaris.core.model.commontypes.MultiLingualContent;
+import rs.teslaris.core.model.commontypes.ResearchArea;
+import rs.teslaris.core.model.user.Authority;
+import rs.teslaris.core.model.user.Privilege;
+import rs.teslaris.core.model.user.User;
+import rs.teslaris.core.model.user.UserNotificationPeriod;
+import rs.teslaris.core.model.user.UserRole;
+import rs.teslaris.core.repository.commontypes.CountryRepository;
+import rs.teslaris.core.repository.commontypes.LanguageRepository;
+import rs.teslaris.core.repository.commontypes.LanguageTagRepository;
+import rs.teslaris.core.repository.commontypes.ResearchAreaRepository;
+import rs.teslaris.core.repository.user.AuthorityRepository;
+import rs.teslaris.core.repository.user.PrivilegeRepository;
+import rs.teslaris.core.repository.user.UserRepository;
+import rs.teslaris.core.util.language.LanguageAbbreviations;
+import rs.teslaris.core.util.seeding.CsvDataLoader;
+import rs.teslaris.core.util.seeding.SKOSLoader;
+
+@Component
+@RequiredArgsConstructor
+@Slf4j
+@Profile("test")
+public class DbInitializer implements ApplicationRunner {
+
+    private final AuthorityRepository authorityRepository;
+
+    private final PrivilegeRepository privilegeRepository;
+
+    private final UserRepository userRepository;
+
+    private final LanguageRepository languageRepository;
+
+    private final LanguageTagRepository languageTagRepository;
+
+    private final PasswordEncoder passwordEncoder;
+
+    private final CountryRepository countryRepository;
+
+    private final ResearchAreaRepository researchAreaRepository;
+
+    private final CsvDataLoader csvDataLoader;
+
+    private final SKOSLoader skosLoader;
+
+    private final Environment environment;
+
+    private final TestingDataInitializer testingDataInitializer;
+
+    private final AssessmentDataInitializer assessmentDataInitializer;
+
+
+    @Override
+    @Transactional
+    public void run(ApplicationArguments args) throws Exception {
+        ///////////////////// NECESSARY DATA /////////////////////
+
+        // PRIVILEGES
+        var allowAccountTakeover = new Privilege("ALLOW_ACCOUNT_TAKEOVER");
+        var takeRoleOfUser = new Privilege("TAKE_ROLE");
+        var deactivateUser = new Privilege("DEACTIVATE_USER");
+        var updateProfile = new Privilege("UPDATE_PROFILE");
+        var createUserBasic = new Privilege("REGISTER_PERSON");
+        var editPersonalInfo = new Privilege("EDIT_PERSON_INFORMATION");
+        var approvePerson = new Privilege("APPROVE_PERSON");
+        var editDocumentFiles = new Privilege("EDIT_DOCUMENT_FILES");
+        var approvePublication = new Privilege("APPROVE_PUBLICATION");
+        var editResearchAreas = new Privilege("EDIT_RESEARCH_AREAS");
+        var editOrganisationUnit = new Privilege("EDIT_ORGANISATION_UNITS");
+        var editOURelations = new Privilege("EDIT_OU_RELATIONS");
+        var editPublishers = new Privilege("EDIT_PUBLISHERS");
+        var editPublicationSeries = new Privilege("EDIT_PUBLICATION_SERIES");
+        var editConferences = new Privilege("EDIT_CONFERENCES");
+        var editEventRelations = new Privilege("EDIT_EVENT_RELATIONS");
+        var mergeJournalPublications = new Privilege("MERGE_JOURNAL_PUBLICATIONS");
+        var mergePersonPublications = new Privilege("MERGE_PERSON_PUBLICATIONS");
+        var mergePersonMetadata = new Privilege("MERGE_PERSON_METADATA");
+        var mergeOUEmployments = new Privilege("MERGE_OU_EMPLOYMENTS");
+        var mergeOUMetadata = new Privilege("MERGE_OU_METADATA");
+        var mergeConferenceProceedings = new Privilege("MERGE_CONFERENCE_PROCEEDINGS");
+        var mergeProceedingsPublications = new Privilege("MERGE_PROCEEDINGS_PUBLICATIONS");
+        var startDeduplicationProcess = new Privilege("START_DEDUPLICATION_PROCESS");
+        var performDeduplication = new Privilege("PERFORM_DEDUPLICATION");
+        var mergeDocumentsMetadata = new Privilege("MERGE_DOCUMENTS_METADATA");
+        var mergeEventMetadata = new Privilege("MERGE_EVENT_METADATA");
+        var mergePublicationSeriesMetadata = new Privilege("MERGE_PUBLICATION_SERIES_METADATA");
+        var mergeMonographPublications = new Privilege("MERGE_MONOGRAPH_PUBLICATIONS");
+        var prepareExportData = new Privilege("PREPARE_EXPORT_DATA");
+        var editIndicators = new Privilege("EDIT_INDICATORS");
+        var editAssessmentClassifications = new Privilege("EDIT_ASSESSMENT_CLASSIFICATIONS");
+        var editAssessmentMeasures = new Privilege("EDIT_ASSESSMENT_MEASURES");
+        var editAssessmentRulebooks = new Privilege("EDIT_ASSESSMENT_RULEBOOKS");
+        var editDocumentIndicators = new Privilege("EDIT_DOCUMENT_INDICATORS");
+        var editCommissions = new Privilege("EDIT_COMMISSIONS");
+        var editEntityIndicatorProofs = new Privilege("EDIT_ENTITY_INDICATOR_PROOFS");
+        var editEntityIndicators = new Privilege("EDIT_ENTITY_INDICATOR");
+        var listMyJournalPublications = new Privilege("LIST_MY_JOURNAL_PUBLICATIONS");
+        var deletePerson = new Privilege("DELETE_PERSON");
+        var registerEmployee = new Privilege("REGISTER_EMPLOYEE");
+        var reindexPrivilege = new Privilege("REINDEX_DATABASE");
+        var mergeBookSeriesPublications = new Privilege("MERGE_BOOK_SERIES_PUBLICATIONS");
+        var editCountries = new Privilege("EDIT_COUNTRIES");
+        var forceDelete = new Privilege("FORCE_DELETE_ENTITIES");
+        var switchEntityToUnmanaged = new Privilege("SWITCH_ENTITY_TO_UNMANAGED");
+        var claimDocument = new Privilege("CLAIM_DOCUMENT");
+        var mergePublisherPublications = new Privilege("MERGE_PUBLISHER_PUBLICATIONS");
+        var mergePublishersMetadata = new Privilege("MERGE_PUBLISHERS_METADATA");
+        var unbindYourselfFromPublication = new Privilege("UNBIND_YOURSELF_FROM_PUBLICATION");
+        var editEntityAssessmentClassifications =
+            new Privilege("EDIT_ENTITY_ASSESSMENT_CLASSIFICATION");
+        var editEventAssessmentClassification =
+            new Privilege("EDIT_EVENT_ASSESSMENT_CLASSIFICATION");
+        var editLanguageTags = new Privilege("EDIT_LANGUAGE_TAGS");
+        var editEventIndicators = new Privilege("EDIT_EVENT_INDICATORS");
+        var scheduleTask = new Privilege("SCHEDULE_TASK");
+        var editPublicationSeriesAssessmentClassifications =
+            new Privilege("EDIT_PUB_SERIES_ASSESSMENT_CLASSIFICATION");
+        var editPubSeriesIndicators = new Privilege("EDIT_PUB_SERIES_INDICATORS");
+        var assessDocument = new Privilege("ASSESS_DOCUMENT");
+
+        privilegeRepository.saveAll(
+            Arrays.asList(allowAccountTakeover, takeRoleOfUser, deactivateUser, updateProfile,
+                createUserBasic, editPersonalInfo, approvePerson, editDocumentFiles,
+                editOrganisationUnit, editResearchAreas, approvePublication, editOURelations,
+                editPublishers, editPublicationSeries, editConferences, editEventRelations,
+                mergeOUEmployments, mergeJournalPublications, mergePersonPublications,
+                mergePersonMetadata, mergeConferenceProceedings, mergeProceedingsPublications,
+                prepareExportData, editIndicators, editAssessmentClassifications, editCommissions,
+                editAssessmentMeasures, editAssessmentRulebooks, editDocumentIndicators,
+                editEntityIndicatorProofs, listMyJournalPublications, deletePerson, scheduleTask,
+                registerEmployee, reindexPrivilege, startDeduplicationProcess, performDeduplication,
+                mergeDocumentsMetadata, mergeEventMetadata, mergePublicationSeriesMetadata,
+                mergeMonographPublications, prepareExportData, mergeBookSeriesPublications,
+                mergeOUMetadata, editCountries, forceDelete, switchEntityToUnmanaged,
+                claimDocument, mergePublisherPublications, mergePublishersMetadata,
+                unbindYourselfFromPublication, editEntityIndicators, editLanguageTags,
+                editEntityAssessmentClassifications, editEventIndicators, editPubSeriesIndicators,
+                editEventAssessmentClassification, editPublicationSeriesAssessmentClassifications,
+                assessDocument));
+
+        // AUTHORITIES
+        var adminAuthority = new Authority(UserRole.ADMIN.toString(), new HashSet<>(
+            List.of(takeRoleOfUser, deactivateUser, updateProfile, editPersonalInfo,
+                createUserBasic, approvePerson, editDocumentFiles, editOrganisationUnit,
+                editResearchAreas, editOURelations, approvePublication, editPublishers,
+                editPublicationSeries, editConferences, editEventRelations, editCommissions,
+                mergeJournalPublications, mergePersonPublications, mergePersonMetadata,
+                mergeOUEmployments, mergeConferenceProceedings, mergeProceedingsPublications,
+                prepareExportData, editIndicators, editAssessmentClassifications, scheduleTask,
+                editAssessmentMeasures, editAssessmentRulebooks, editDocumentIndicators,
+                editEntityIndicatorProofs, deletePerson, registerEmployee, reindexPrivilege,
+                startDeduplicationProcess, performDeduplication, mergeDocumentsMetadata,
+                mergeEventMetadata, mergePublicationSeriesMetadata, mergeMonographPublications,
+                prepareExportData, mergeBookSeriesPublications, mergeOUMetadata, editCountries,
+                forceDelete, switchEntityToUnmanaged, mergePublisherPublications, editLanguageTags,
+                mergePublishersMetadata, editEntityIndicators, editEntityAssessmentClassifications,
+                editEventIndicators, editEventAssessmentClassification, editPubSeriesIndicators,
+                editPublicationSeriesAssessmentClassifications, assessDocument
+            )));
+
+        var researcherAuthority = new Authority(UserRole.RESEARCHER.toString(), new HashSet<>(
+            List.of(allowAccountTakeover, updateProfile, editPersonalInfo, assessDocument,
+                createUserBasic, editDocumentFiles, editDocumentIndicators, claimDocument,
+                editEntityIndicatorProofs, listMyJournalPublications,
+                unbindYourselfFromPublication, editEntityIndicators)));
+
+        var institutionalEditorAuthority =
+            new Authority(UserRole.INSTITUTIONAL_EDITOR.toString(), new HashSet<>(
+                List.of(new Privilege[] {updateProfile, allowAccountTakeover})));
+
+        var commissionAuthority =
+            new Authority(UserRole.COMMISSION.toString(), new HashSet<>(List.of(
+                editEventAssessmentClassification, updateProfile, editEventIndicators,
+                editPublicationSeriesAssessmentClassifications, editPubSeriesIndicators,
+                allowAccountTakeover, editEntityIndicatorProofs
+            )));
+
+        authorityRepository.saveAll(
+            List.of(adminAuthority, researcherAuthority, institutionalEditorAuthority,
+                commissionAuthority));
+
+        // LANGUAGES
+        var serbianLanguage = new Language();
+        serbianLanguage.setLanguageCode(LanguageAbbreviations.SERBIAN);
+        var englishLanguage = new Language();
+        englishLanguage.setLanguageCode(LanguageAbbreviations.ENGLISH);
+        // We will maybe need YU, I don't know
+        var yuLanguage = new Language();
+        yuLanguage.setLanguageCode(LanguageAbbreviations.YUGOSLAV);
+        yuLanguage.setDeleted(true);
+        var germanLanguage = new Language();
+        germanLanguage.setLanguageCode(LanguageAbbreviations.GERMAN);
+        var frenchLanguage = new Language();
+        frenchLanguage.setLanguageCode(LanguageAbbreviations.FRENCH);
+        var spanishLanguage = new Language();
+        spanishLanguage.setLanguageCode(LanguageAbbreviations.SPANISH);
+        var russianLanguage = new Language();
+        russianLanguage.setLanguageCode(LanguageAbbreviations.RUSSIAN);
+        var croatianLanguage = new Language();
+        croatianLanguage.setLanguageCode(LanguageAbbreviations.CROATIAN);
+        var italianLanguage = new Language();
+        italianLanguage.setLanguageCode(LanguageAbbreviations.ITALIAN);
+        var slovenianLanguage = new Language();
+        slovenianLanguage.setLanguageCode(LanguageAbbreviations.SLOVENIAN);
+
+        languageRepository.saveAll(
+            List.of(serbianLanguage, englishLanguage, yuLanguage, germanLanguage, frenchLanguage,
+                spanishLanguage, russianLanguage, croatianLanguage, italianLanguage,
+                slovenianLanguage));
+
+        // LANGUAGE TAGS
+        var englishTag = new LanguageTag(LanguageAbbreviations.ENGLISH, "English");
+        var serbianTag = new LanguageTag(LanguageAbbreviations.SERBIAN, "Srpski");
+        var hungarianTag = new LanguageTag(LanguageAbbreviations.HUNGARIAN, "Magyar");
+        var germanTag = new LanguageTag(LanguageAbbreviations.GERMAN, "Deutsch");
+        var frenchTag = new LanguageTag(LanguageAbbreviations.FRENCH, "Français");
+        var spanishTag = new LanguageTag(LanguageAbbreviations.SPANISH, "Español");
+        var russianTag = new LanguageTag(LanguageAbbreviations.RUSSIAN, "Русский");
+        var croatianTag = new LanguageTag(LanguageAbbreviations.CROATIAN, "Croatian");
+        var italianTag = new LanguageTag(LanguageAbbreviations.ITALIAN, "Italian");
+        var slovenianTag = new LanguageTag(LanguageAbbreviations.SLOVENIAN, "Slovenian");
+        languageTagRepository.saveAll(
+            List.of(englishTag, serbianTag, hungarianTag, germanTag, frenchTag, spanishTag,
+                russianTag, croatianTag, italianTag, slovenianTag));
+
+        // ADMIN USER
+        var adminUser =
+            new User("admin@admin.com", passwordEncoder.encode("admin"), "note", "Marko",
+                "Markovic", false, false, serbianLanguage, adminAuthority, null, null, null,
+                UserNotificationPeriod.DAILY);
+        userRepository.save(adminUser);
+
+        // RESEARCH AREAS - NOT COMPLETE
+        var researchArea1 = new ResearchArea(new HashSet<>(Set.of(
+            new MultiLingualContent(serbianTag, "Elektrotehnicko i racunarsko inzenjerstvo", 2))),
+            new HashSet<>(), null);
+        var researchArea2 = new ResearchArea(new HashSet<>(Set.of(
+            new MultiLingualContent(serbianTag, "Softversko inzenjerstvo", 2))),
+            new HashSet<>(), researchArea1);
+        var researchArea3 = new ResearchArea(new HashSet<>(Set.of(
+            new MultiLingualContent(serbianTag, "Cybersecurity", 2))),
+            new HashSet<>(), researchArea2);
+
+        researchAreaRepository.saveAll(List.of(researchArea1, researchArea2, researchArea3));
+
+        // COUNTRIES
+        csvDataLoader.loadData("countries.csv", this::processCountryLine, ',');
+
+        // RESEARCH AREAS
+        skosLoader.loadResearchAreas();
+
+        ///////////////////// ASSESSMENT DATA /////////////////////
+        assessmentDataInitializer.initializeIndicators(englishTag, serbianTag);
+        assessmentDataInitializer.initializeAssessmentClassifications(englishTag, serbianTag);
+        var commission5 = assessmentDataInitializer.initializeCommissions(englishTag, serbianTag);
+
+        ///////////////////// TESTING DATA /////////////////////
+        if (Arrays.stream(environment.getActiveProfiles())
+            .anyMatch(profile -> profile.equalsIgnoreCase("test"))) {
+            testingDataInitializer.initializeIntegrationTestingData(serbianTag, serbianLanguage,
+                englishTag,
+                germanLanguage, researchArea3, researcherAuthority, commissionAuthority,
+                commission5);
+        }
+    }
+
+    private void processCountryLine(String[] line) {
+        var country = new Country();
+        country.setCode(line[0]);
+
+        var names = new HashSet<MultiLingualContent>();
+        var nameList = line[1].split(";");
+        for (int i = 0; i < nameList.length; i++) {
+            var nameParts = nameList[i].split("@");
+            if (nameParts.length == 2) {
+                var content = nameParts[0].trim();
+                var languageTagCode = nameParts[1].trim().toUpperCase();
+
+                var languageTag = languageTagRepository
+                    .findLanguageTagByLanguageTag(languageTagCode)
+                    .orElseGet(() -> {
+                        var newLanguageTag = new LanguageTag();
+                        newLanguageTag.setLanguageTag(languageTagCode);
+                        newLanguageTag.setDisplay(languageTagCode);
+                        log.info("Created new language tag: {}", languageTagCode);
+                        return languageTagRepository.save(newLanguageTag);
+                    });
+
+                var multilingualContent =
+                    new MultiLingualContent(languageTag, content, i + 1);
+                names.add(multilingualContent);
+            }
+        }
+        country.setName(names);
+        countryRepository.save(country);
+    }
+}
